requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
clickhouse-connect==0.14.1
idna==3.11
lz4==4.4.5
polars==1.39.0
polars-runtime-32==1.39.0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
urllib3==2.6.3
zstandard==0.25.0
